Updated ore generation and mushroom trees
Changed ore generation to make actual veins of ores and not spots of
them.
Changed mushroom trees to be generated with own code instead of with the
supplied method, to stop half being cut-off.
Moved the island data generation code into it's own function, as it was
used in multiple areas.