Configurable island changes
Added config values to allow changing the chances of each island type,
these values are used in a 'weight' system. The bigger the value
compared to the rest, the more likely it is.
The spawn point of the world will be set to an island which should have
trees if it detects that the current spawn is invalid (highest Y at the
spawn x/z is at least 3 blocks below it).